pyproject.toml
[tool.poetry]
name = "geoparser"
version = "0.1.0"
description = ""
authors = ["AndreyKosov <andrey.kosov.05@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
bs4 = "^0.0.2"
fastapi = {extras = ["all"], version = "^0.115.6"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
